backend/db/migrations/env.py
"""Alembic environment — reads DATABASE_URL from .env, uses sync psycopg2 driver."""

import os
import re
import sys
from logging.config import fileConfig
from pathlib import Path

from alembic import context
from sqlalchemy import engine_from_config, pool

# Make sure `backend/` is on sys.path so our modules are importable
sys.path.insert(0, str(Path(__file__).resolve().parents[2]))

# Load .env so DATABASE_URL is available when running alembic from the shell
from dotenv import load_dotenv

load_dotenv(Path(__file__).resolve().parents[2] / ".env")

# Import Base + all models so autogenerate can see every table
from db.database import Base  # noqa: E402
import db.models  # noqa: F401, E402  registers Syllabus + Chunk on Base.metadata

config = context.config

if config.config_file_name is not None:
    fileConfig(config.config_file_name)

target_metadata = Base.metadata


def _sync_url(url: str) -> str:
    """Convert asyncpg URL to psycopg2 URL for Alembic's sync engine."""
    return re.sub(r"postgresql\+asyncpg://", "postgresql+psycopg2://", url)


def _get_url() -> str:
    url = os.environ.get("DATABASE_URL", "")
    if not url:
        raise RuntimeError("DATABASE_URL environment variable is not set.")
    return _sync_url(url)


def run_migrations_offline() -> None:
    """Generate SQL script without a live DB connection."""
    context.configure(
        url=_get_url(),
        target_metadata=target_metadata,
        literal_binds=True,
        dialect_opts={"paramstyle": "named"},
    )
    with context.begin_transaction():
        context.run_migrations()


def run_migrations_online() -> None:
    """Apply migrations against a live DB connection."""
    cfg = config.get_section(config.config_ini_section, {})
    cfg["sqlalchemy.url"] = _get_url()

    connectable = engine_from_config(cfg, prefix="sqlalchemy.", poolclass=pool.NullPool)

    with connectable.connect() as connection:
        context.configure(connection=connection, target_metadata=target_metadata)
        with context.begin_transaction():
            context.run_migrations()


if context.is_offline_mode():
    run_migrations_offline()
else:
    run_migrations_online()
