backend/requirements.txt
# FastAPI + server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Config / env
python-dotenv==1.0.1
pydantic-settings==2.7.0

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.37
alembic==1.14.0
pgvector==0.3.6
psycopg2-binary==2.9.10

# File parsing
pymupdf==1.25.3

# LLM / AI
anthropic==0.45.0
openai==1.63.0

# Embeddings / reranking
sentence-transformers==3.4.1

# Utilities
httpx==0.28.1
